Add files via upload
The project was implemented using Python with libraries such as scikit-learn for machine learning, BeautifulSoup & requests for web scraping, Tkinter for GUI, and pandas/numpy for data handling. A dataset of around 5,000 labeled samples was preprocessed, vectorized using TF-IDF, and used to train models like Logistic Regression, SVM, and Naive Bayes. The system was designed to scrape website text, analyze it, classify it as benign or containing dark patterns, and present results through a user-friendly GUI.